More general PredictionModel and TrainableModel
Remove reliance upon OpenCVStatModel in favor of permitting a much simpler PredictionModel to be used.
