+++ qupath-extension-processing/src/main/java/qupath/process/gui/commands/ObjectClassifierCommand.java
@@ -4,7 +4,7 @@
  * %%
  * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
  * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2025 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2018 - 2026 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -99,7 +99,8 @@ import qupath.lib.objects.hierarchy.events.PathObjectHierarchyEvent;
 import qupath.lib.objects.hierarchy.events.PathObjectHierarchyListener;
 import qupath.lib.projects.ProjectImageEntry;
 import qupath.opencv.ml.models.OpenCVClassifiers;
-import qupath.opencv.ml.models.OpenCVStatModel;
+import qupath.opencv.ml.models.PredictionModel;
+import qupath.opencv.ml.models.TrainableModel;
 import qupath.opencv.ml.models.RTreesClassifier;
 import qupath.opencv.ml.objects.OpenCVMLClassifier;
 import qupath.opencv.ml.objects.features.FeatureExtractor;
@@ -286,7 +287,7 @@ public class ObjectClassifierCommand implements Runnable {
 
 		private ReadOnlyObjectProperty<PathObjectFilter> objectFilter;
 
-		private ReadOnlyObjectProperty<OpenCVStatModel> selectedModel;
+		private ReadOnlyObjectProperty<TrainableModel> selectedModel;
 
 		private ReadOnlyObjectProperty<OutputClasses> outputClasses;
 		private ReadOnlyObjectProperty<TrainingFeatures> trainingFeatures;
@@ -495,7 +496,7 @@ public class ObjectClassifierCommand implements Runnable {
 		 */
 		private FutureTask<ObjectClassifier<BufferedImage>> createClassifierUpdateTask(boolean doClassification) {
 			var filter = objectFilter.get();
-			OpenCVStatModel statModel = selectedModel == null ? null : selectedModel.get();
+			TrainableModel statModel = selectedModel == null ? null : selectedModel.get();
 			if (statModel == null) {
 				logger.warn("No classifier - cannot update classifier");
 				resetPieChart();
@@ -747,7 +748,7 @@ public class ObjectClassifierCommand implements Runnable {
 		private static ObjectClassifier<BufferedImage> createClassifier(
 				Collection<TrainingData<BufferedImage>> training,
 				PathObjectFilter filter,
-				OpenCVStatModel statModel,
+				TrainableModel statModel,
 				FeatureExtractor<BufferedImage> extractor,
 				Normalization normalization,
 				double pcaRetainedVariance,
@@ -856,7 +857,7 @@ public class ObjectClassifierCommand implements Runnable {
 		 */
 		@SuppressWarnings("unchecked")
 		private static <T> FeatureExtractor<T> updateFeatureExtractorAndTrainClassifier(
-				OpenCVStatModel classifier,
+				TrainableModel classifier,
 				Collection<TrainingData<T>> trainingCollection,
 				FeatureExtractor<T> extractor,
 				Normalization normalization,
@@ -994,7 +995,7 @@ public class ObjectClassifierCommand implements Runnable {
 			return extractor;
 		}
 
-		static boolean trainClassifier(OpenCVStatModel classifier, Mat matFeatures, Mat matTargets, int nLabels, boolean doMulticlass) {
+		static boolean trainClassifier(TrainableModel classifier, Mat matFeatures, Mat matTargets, int nLabels, boolean doMulticlass) {
 			// Train classifier
 			// TODO: Optionally limit the number of training samples we use
 			long startTime = System.currentTimeMillis();
@@ -1152,7 +1153,7 @@ public class ObjectClassifierCommand implements Runnable {
 			 * Classifier type
 			 */
 			var labelClassifier = new Label("Classifier");
-			var comboClassifier = new ComboBox<OpenCVStatModel>();
+			var comboClassifier = new ComboBox<TrainableModel>();
 			comboClassifier.getItems().addAll(
 					OpenCVClassifiers.createStatModel(RTrees.class),
 					OpenCVClassifiers.createStatModel(ANN_MLP.class),
